+++ lib/cldr/number/format.ex
@@ -442,8 +442,24 @@ defmodule Cldr.Number.Format do
   ## Example
 
       iex> Cldr.Number.Format.format_styles_for("en", :latn, MyApp.Cldr)
-      {:ok, [:accounting, :currency, :currency_long, :currency_short,
-      :decimal_long, :decimal_short, :percent, :scientific, :standard]}
+      {
+        :ok,
+        [
+          :accounting,
+          :accounting_alpha_next_to_number,
+          :accounting_no_symbol,
+          :currency,
+          :currency_alpha_next_to_number,
+          :currency_long,
+          :currency_no_symbol,
+          :currency_short,
+          :decimal_long,
+          :decimal_short,
+          :percent,
+          :scientific,
+          :standard
+        ]
+      }
 
   """
   @reject_styles [:__struct__, :currency_spacing, :other]
@@ -542,8 +558,21 @@ defmodule Cldr.Number.Format do
   ## Example
 
       iex> Cldr.Number.Format.decimal_format_styles_for("en", :latn, MyApp.Cldr)
-      {:ok, [:accounting, :currency, :currency_long, :percent,
-       :scientific, :standard]}
+      {
+        :ok,
+        [
+          :accounting,
+          :accounting_alpha_next_to_number,
+          :accounting_no_symbol,
+          :currency,
+          :currency_alpha_next_to_number,
+          :currency_long,
+          :currency_no_symbol,
+          :percent,
+          :scientific,
+          :standard
+        ]
+      }
 
   """
   @spec decimal_format_styles_for(