+++ test/support/number_format_test_data.exs
@@ -156,9 +156,9 @@ defmodule Cldr.Test.Number.Format do
       {123.4, "123.40 $", [format: "#,##0.00 ¤", locale: "en-AU-u-cu-aud"]},
       {123.4, "123.40 USD", [format: "#,##0.00 ¤", locale: "en-AU-u-cu-aud", currency: :USD]},
 
-      # Specify accounting or currency in the locale
-      {-123.4, "$-123.40", [format: :currency, locale: "en-AU-u-cu-aud-cf-standard"]},
-      {-123.4, "($123.40)", [format: :currency, locale: "en-AU-u-cu-aud-cf-account"]},
+      # Specify accounting or currency in the locale doesn't change the format
+      {-123.4, "-123.40", [format: :currency, locale: "en-AU-u-cu-aud-cf-standard"]},
+      {-123.4, "-123.40", [format: :currency, locale: "en-AU-u-cu-aud-cf-account"]},
 
       # Formatting digital tokens (Crypto currencies)
       {1234545656.456789, "₿ 1,234,545,656.456789", [currency: "BTC"]},